Refactor modules that depend on aind services
