+++ src/DataSchemas/aind_behavior_services/aind_services/watchdog.py
@@ -1,6 +1,3 @@
-from os import PathLike
-from typing import Optional
-
 try:
     import aind_watchdog_service  # noqa: F401
 except ImportError as e:
@@ -11,12 +8,17 @@ except ImportError as e:
     )
     raise
 
-from aind_watchdog_service.models.manifest_config import ManifestConfig
+import os
+import datetime
+from pathlib import Path
+from typing import Optional
+
+from aind_watchdog_service.models.manifest_config import ManifestConfig, Platform
 from aind_watchdog_service.models.watch_config import WatchConfig
 
 
 def create_watchdog_config(
-    watched_directory: PathLike, manifest_complete_directory: PathLike, webhook_url: Optional[str] = None
+    watched_directory: os.PathLike, manifest_complete_directory: os.PathLike, webhook_url: Optional[str] = None
 ) -> WatchConfig:
     """Create a WatchConfig object"""
     return WatchConfig(
@@ -24,3 +26,4 @@ def create_watchdog_config(
         manifest_complete=str(manifest_complete_directory),
         webhook_url=webhook_url,
     )
+
